perf(charts): lazy-load the zoom dialog so charts stop shipping the data-table
ChartContainer wraps every chart in the app and statically imported
ChartZoomDialog, which statically imports the whole data-table system
(~11.4k lines). Every page with a chart shipped the full table stack for a
dialog that only opens on a deliberate click.

The render site was already gated on `zoomOpen` — only the import was wrong.
Third instance of the same failure mode as #2995 and #2997: a boundary that
looks deferred but isn't, because the expensive module is reachable from code
that always runs.

ChartZoomButton lived in the same module, so importing just the button also
pulled the dialog and everything under it. It moves to its own file and
chart-zoom-dialog re-exports it, so no call site changes.

Local production build, initial load:

  /overview  253 scripts / 2826KB, data-table loaded
          -> 238 scripts / 2410KB, data-table absent      -416KB
  /merges    237 scripts / 2607KB, data-table still loaded (correct)

Suspense uses fallback={null} so the card stays put and the dialog appears
once its chunk lands, rather than flashing a skeleton over the page.

Closes #3000
